Reorganize Osmosis gamm parser package for easier development
